requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14
psycopg2-binary>=2.9
google-generativeai>=0.3.2
python-dotenv>=1.0.0
